Add comment for new page change behavior
Previously using the PDFPageView and changing the page there resulted in
a just the pdf page get changed out with the new one.
By using the PDFSInglePageViewer a page change results in focusing the new
pdf page requesting focus "snapping" into view.
If the page change is triggered by an input element whose position is fixed,
it may get out of the users viewport, which isn't very pretty. The input still
receives inputs (like pressing up/down keys on keyboard) but the user can't
see the input element itself (an example of this behavior can be found in
by executing "ng serve".